% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bareun.R
\name{analyze_text}
\alias{analyze_text}
\title{analyze_text}
\usage{
analyze_text(tagged, text)
}
\arguments{
\item{tagged}{Bareun tagger result}

\item{text}{string - input text}
}
\value{
raw result
}
\description{
analyze_text
}
