% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bareun.R
\name{as_json_string}
\alias{as_json_string}
\title{Return JSON string for response message}
\usage{
as_json_string(tagged, pretty = FALSE)
}
\arguments{
\item{tagged}{Bareun tagger result}

\item{pretty}{bool - TRUE 면 들여쓰기해서 읽기 좋게 만든다}
}
\value{
returns JSON string
}
\description{
\itemize{
\item 결과를 JSON 문자열로 출력
}
}
