% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bareun.R
\name{build_dict_set}
\alias{build_dict_set}
\title{Build A Dictionary}
\usage{
build_dict_set(tagged, domain, name, dict_set)
}
\arguments{
\item{tagged}{BareunNLP tagger result}

\item{domain}{domain name of custom dictionary}

\item{name}{name of dictionary set}

\item{dict_set}{set of dictionary contents(values)}
}
\value{
returns DictSet
}
\description{
\itemize{
\item 사용자 사전 한 세트 만들기
}
}
