% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bareun.R
\name{dict_list}
\alias{dict_list}
\title{Get List of Custom Dictionaries}
\usage{
dict_list(tagged)
}
\arguments{
\item{tagged}{Bareun tagger result}
}
\value{
returns dict
}
\description{
\itemize{
\item 사용자 사전의 목록 출력
}
}
