% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict-search.R
\name{dict_words}
\alias{dict_words}
\title{Get matched words as a character vector}
\usage{
dict_words(result)
}
\arguments{
\item{result}{search_dict() 결과}
}
\value{
returns character vector of words
}
\description{
\itemize{
\item 검색 결과에서 표제어만 꺼낸다
}
}
