% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{.bareun_stop}
\alias{.bareun_stop}
\title{조건 클래스를 붙여 오류를 던진다.}
\usage{
.bareun_stop(class, message, ...)
}
\arguments{
\item{class}{character - 조건 클래스들. 좁은 것부터 넓은 것 순으로 준다.}

\item{message}{string - 사람이 읽는 오류 메시지}

\item{...}{조건 객체에 함께 실을 값들}
}
\description{
조건 클래스를 붙여 오류를 던진다.
}
\keyword{internal}
