% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{.m}
\alias{.m}
\title{메시지를 꺼내 인자를 채운다.}
\usage{
.m(key, ...)
}
\arguments{
\item{key}{string - .messages 의 키}

\item{...}{sprintf 로 채울 값들}
}
\value{
완성된 메시지 문자열
}
\description{
메시지를 꺼내 인자를 채운다.
}
\keyword{internal}
