% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bareun.R
\name{get_dict}
\alias{get_dict}
\title{Get Custom Dictionary}
\usage{
get_dict(tagged, name)
}
\arguments{
\item{tagged}{BareunNLP tagger result}

\item{name}{name of custom dictionary}
}
\value{
returns dict
}
\description{
\itemize{
\item 지정된 사용자 사전 읽어오기
}
}
