% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bareun.R
\name{get_set}
\alias{get_set}
\title{Get Contents of Set}
\usage{
get_set(tagged, set_name)
}
\arguments{
\item{tagged}{BareunNLP tagger result}

\item{set_name}{name of set (np, cp, caret)}
}
\value{
returns list of words
}
\description{
사용자 사전 세트별 내용 출력
}
