% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bareun.R
\name{make_custom_dict}
\alias{make_custom_dict}
\title{Update Custom Dictionary}
\usage{
make_custom_dict(tagged, domain, nps, cps, carets, vvs, vas)
}
\arguments{
\item{tagged}{Bareun tagger result}

\item{domain}{domain name of custom dictionary}

\item{nps}{set of np-set dictinary}

\item{cps}{set of cp-set dictinary}

\item{carets}{set of cp-caret-set dictinary}

\item{vvs}{set of vv-set dictinary}

\item{vas}{set of va-set dictinary}
}
\value{
print result
}
\description{
\itemize{
\item 사용자 사전 만들고 업로드(저장)
}
}
