% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bareun.R
\name{nouns}
\alias{nouns}
\title{Return array of Nouns}
\usage{
nouns(tagged = NULL, text = "")
}
\arguments{
\item{tagged}{Bareun tagger result}

\item{text}{input text}
}
\value{
returns array of list for nouns
}
\description{
\itemize{
\item 결과/문장의 명사 리스트만 출력
\item 새로운 문장이 주어지면 결과를 변경하고, 문장이 주어지지 않으면 이전 결과를 다시 사용
}
}
\examples{
\dontrun{
nouns(text = "결과를 문자열로 바꾼다.")
}

# 결과 예시:
# [[1]]
# [1] "결과" "문자열"
}
