% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bareun.R
\name{pos}
\alias{pos}
\title{Return array of morpheme/postag words}
\usage{
pos(tagged = NULL, text = "")
}
\arguments{
\item{tagged}{Bareun tagger result}

\item{text}{input text}
}
\value{
returns array of words 'morpheme/tag'
}
\description{
\itemize{
\item 결과/문장을 '음절/태그' 문자열 리스트로 출력
\item 새로운 문장이 주어지면 결과를 변경하고, 문장이 주어지지 않으면 이전 결과를 다시 사용
}
}
\examples{
\dontrun{
pos(text = "결과를 문자열로 바꾼다.")
}

# 결과 예시:
# [[1]]
# [1] "결과/NNG" "를/JKO" "문자열/NNG" "로/JKB" "바꾸/VV" "ㄴ다/EF" "./SF"
}
