% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bareun.R
\name{postag}
\alias{postag}
\title{Return array of (morpheme, postag) pairs}
\usage{
postag(tagged = NULL, text = "", matrix = FALSE)
}
\arguments{
\item{tagged}{Bareun tagger result}

\item{text}{input text}

\item{matrix}{if TRUE, result output to matrix not list (default = FALSE)}
}
\value{
returns array of lists for (morpheme, tag)
}
\description{
\itemize{
\item 결과/문장을 (음절, 형태소태그) 리스트의 리스트로 출력
\item 새로운 문장이 주어지면 결과를 변경하고, 문장이 주어지지 않으면 이전 결과를 다시 사용
}
}
\examples{
\dontrun{
postag(text = "결과를 문자열로 바꾼다.", matrix = TRUE)
}

# 결과 예시:
# [[1]]
#      [,1]     [,2]
# [1,] "결과"   "NNG"
# [2,] "를"     "JKO"
# [3,] "문자열" "NNG"
# [4,] "로"     "JKB"
# [5,] "바꾸"   "VV"
# [6,] "ㄴ다"   "EF"
# [7,] "."      "SF"
}
