% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bareun.R
\name{print_as_json}
\alias{print_as_json}
\title{Print JSON string for response message}
\usage{
print_as_json(tagged)
}
\arguments{
\item{tagged}{BareunNLP tagger result}
}
\value{
prints JSON string
}
\description{
결과를 JSON 문자열로 출력.
}
