% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bareun.R
\name{print_dict_all}
\alias{print_dict_all}
\title{Print All Contents of Custom Dictionary}
\usage{
print_dict_all(tagged)
}
\arguments{
\item{tagged}{BareunNLP tagger result}
}
\value{
prints all contents of all sets
}
\description{
사용자 사전 내용 모두 출력
}
