% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision.R
\name{revised_text}
\alias{revised_text}
\title{Get the corrected sentence}
\usage{
revised_text(revised)
}
\arguments{
\item{revised}{correct_grammar() 결과}
}
\value{
returns corrected text as string
}
\description{
\itemize{
\item 교정된 문장만 꺼낸다
}
}
