% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision.R
\name{revisions}
\alias{revisions}
\title{Get corrections as a data frame}
\usage{
revisions(revised)
}
\arguments{
\item{revised}{correct_grammar() 결과}
}
\value{
returns data.frame with columns origin, revised, category, help
}
\description{
\itemize{
\item 무엇이 어떻게 바뀌었는지 표로 만든다
}
}
