% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bareun.R
\name{as_json_string}
\alias{as_json_string}
\title{Return JSON string for response message}
\usage{
as_json_string(tagged)
}
\arguments{
\item{tagged}{BareunNLP tagger result}
}
\value{
returns JSON string
}
\description{
\itemize{
\item 결과를 JSON 문자열로 출력
}
}
