% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bareun.R
\name{morphs}
\alias{morphs}
\title{Return array of Morphemes}
\usage{
morphs(tagged = NULL, text = "")
}
\arguments{
\item{tagged}{BareunNLP tagger result}

\item{text}{input text}
}
\value{
returns array of list for morphemes
}
\description{
\itemize{
\item 결과/문장의 음절 리스트만 출력
\item 새로운 문장이 주어지면 결과를 변경하고, 문장이 주어지지 않으면 이전 결과를 다시 사용
}
}
\examples{
> morphs(, "결과를 문자열로 바꾼다.")
[[1]]
[1] "결과"   "를"     "문자열" "로"     "바꾸"   "ㄴ다"   "."
}
