% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bareun.R
\name{postag}
\alias{postag}
\title{Return array of (morpheme, postag) pairs}
\usage{
postag(tagged = NULL, text = "", matrix = FALSE)
}
\arguments{
\item{tagged}{BareunNLP tagger result}

\item{matrix}{if TRUE, result output to matrix not list (default = FALSE)}
}
\value{
returns array of lists for (morpheme, tag)
}
\description{
결과에서 (음절, 형태소 태그)의 배열을 반환.
}
\examples{
> postag(, "결과를 문자열로 바꾼다.", TRUE)
[[1]]
     [,1]     [,2]
[1,] "결과"   "NNG"
[2,] "를"     "JKO"
[3,] "문자열" "NNG"
[4,] "로"     "JKB"
[5,] "바꾸"   "VV"
[6,] "ㄴ다"   "EF"
[7,] "."      "SF"
}
