% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bareun.R
\name{print_as_json}
\alias{print_as_json}
\title{Print JSON string for response message}
\usage{
print_as_json(tagged)
}
\arguments{
\item{tagged}{Bareun tagger result}
}
\value{
prints JSON string
}
\description{
\itemize{
\item 결과를 읽을 수 있는 JSON 문자열로 화면 출력
}
}
