% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bareun.R
\name{remove_custom_dict}
\alias{remove_custom_dict}
\title{Remove Custom Dictionary}
\usage{
remove_custom_dict(tagged, names)
}
\arguments{
\item{tagged}{BareunNLP tagger result}

\item{name}{name of custom dictionary}
}
\value{
print results
}
\description{
\itemize{
\item 사용자 사전(들)을 삭제
}
}
