% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bareun.R
\name{set_api}
\alias{set_api}
\title{save api config}
\usage{
set_api(apikey, server = "localhost", port = 5757, api = "rest")
}
\arguments{
\item{apikey}{string - bareun user's api key}

\item{server}{string - bareun api server name or ip}

\item{port}{number - bareun api server port (default: 5757)}

\item{api}{string - api type (rest or grpc)}
}
\description{
save api config
}
