% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bareun.R
\name{set_api}
\alias{set_api}
\title{save api config}
\usage{
set_api(apikey, host = "localhost:5757", api = "grpc")
}
\arguments{
\item{apikey}{string - bareun user's api key}

\item{host}{string - bareun api server (addr:port)}

\item{api}{string - api type (grpc or rest)}
}
\description{
save api config
}
