% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bareun.R
\name{set_server}
\alias{set_server}
\title{save server config}
\usage{
set_server(host = "localhost:5757", api = "rest")
}
\arguments{
\item{host}{string - bareun api server (addr:port)}

\item{api}{string - api type (grpc or rest)}
}
\description{
save server config
}
