% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bareun.R
\name{tagger}
\alias{tagger}
\title{Call Bareun server to read postag result message for the sentences}
\usage{
tagger(
  text = "",
  apikey = "",
  server = "",
  port = 5656,
  domain = "",
  local = FALSE,
  bareun = TRUE,
  api = ""
)
}
\arguments{
\item{text}{string - subject sentences splitted by newlines(\\n)}

\item{apikey}{string - Bareun user's API KEY}

\item{server}{string - Bareun server address}

\item{port}{number - Bareun server port}

\item{domain}{string - custom domain (custom dictionary)}

\item{local}{bool - use local protobuf files, if TRUE}
}
\value{
returns tagged object
}
\description{
\itemize{
\item Bareun 서버를 호출하여 입력 문장(들)의 분석 결과를 가져 온다
}
}
\examples{
tagged <- tagger("결과를 문자열로 바꾼다.")
}
