% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bareun.R
\name{tagger}
\alias{tagger}
\title{Call BareunNLP server to read postag result message for the sentences}
\usage{
tagger(
  text = "",
  server = "nlp.baikal.ai",
  port = 5656,
  domain = "",
  local = FALSE
)
}
\arguments{
\item{text}{string - subject sentences splitted by newlines(\\n)}

\item{server}{string - BareunNLP grpc server address}

\item{port}{number - BareunNLP grpc server port}

\item{domain}{string - custom domain (custom dictionary)}

\item{local}{bool - use local protobuf files, if TRUE}
}
\value{
returns tagged object
}
\description{
\itemize{
\item BareunNLP grpc 서버를 호출하여 입력 문장(들)의 분석 결과를 가져 온다
}
}
\examples{
tagged <- tagger("결과를 문자열로 바꾼다.")
}
