% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bareun.R
\name{verbs}
\alias{verbs}
\title{Return array of Verbs}
\usage{
verbs(tagged = NULL, text = "")
}
\arguments{
\item{tagged}{BareunNLP tagger result}

\item{text}{input text}
}
\value{
returns array of list for verbs
}
\description{
\itemize{
\item 결과/문장의 동사 리스트만 출력
\item 새로운 문장이 주어지면 결과를 변경하고, 문장이 주어지지 않으면 이전 결과를 다시 사용
}
}
\examples{
> verbs(, "결과를 문자열로 바꾼다.")
[[1]]
[1] "바꾸"
}
