% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bareun.R
\name{verbs}
\alias{verbs}
\title{Return array of Verbs}
\usage{
verbs(tagged = NULL, text = "")
}
\arguments{
\item{tagged}{BareunNLP tagger result}
}
\value{
returns array of list for verbs
}
\description{
형태소 분석 결과의 동사 배열 반환.
}
\examples{
> verbs(, "결과를 문자열로 바꾼다.")
[[1]]
[1] "바꾸"
}
